hub-mirror/requirements.txt
PyYAML>=6.0.1
GitPython==3.1.13
requests==2.25.1
tenacity==6.3.1
click==8.1.7
